pyproject.toml
[tool.poetry]
name = "proposal-droid"
version = "0.1.0"
description = ""
authors = ["cem"]
readme = "README.md"
packages = [{include = "proposal_droid"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "0.28"
requests = "^2.32.3"
python-docx = "^1.1.2"
opencv-python = "^4.10.0.84"
numpy = "^2.0.0"
torch = "^2.3.1"
torchaudio = "^2.3.1"
transformers = "^4.42.3"
pillow = "^10.4.0"
openpyxl = "^3.1.5"
pytesseract = "^0.3.10"
pandas = "^2.2.2"
tesseract-ocr = "^0.0.1"
pyheif = "^0.7.1"
docx = "^0.2.4"
soundfile = "^0.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
